F3.5-fix3-b: الإصلاح الفعلي لم يُكتب سابقاً (خطأ cwd جعل الكومِت السابق يحمل التوثيق حصراً) — الآن: double.clamp() يرجع num، toDouble() إلزامية

+++ app/lib/features/lab/spring_lab_screen.dart
@@ -122,7 +122,8 @@ class _SpringLabScreenState extends State<SpringLabScreen>
 
   void _dragTo(double normalized) {
     // سحب المستخدم ⇒ الشروط الابتدائية (مطال + بداية ساكنة)
-    final clamped = normalized.clamp(-1.0, 1.0);
+    // ⚠️ double.clamp يرجع num — .toDouble() إلزامي قبل العمر double
+    final clamped = normalized.clamp(-1.0, 1.0).toDouble();
     _ticker.stop();
     setState(() {
       _running = false;
